@@ -2,7 +2,9 @@
 
 namespace ByJG\AnyDataset\Db\Exception;
 
-class TransactionStartedException extends \Exception
+use Exception;
+
+class TransactionStartedException extends Exception
 {
 
 }
